Create Task try 1

+++ oci-react-samples/MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/repository/TaskUpdateRepository.java
@@ -12,6 +12,6 @@ import javax.transaction.Transactional;
 @Transactional
 @EnableTransactionManagement
 
-public interface TaskUpdateRepository extends JpaRepository<TaskUpdate,Integer> {
+public interface TaskUpdateRepository extends JpaRepository<TaskUpdate, Long> {
 
 }
